Automatically encode as iso-8859-1 if possible
LinkMobility (PsWin) will use the GSM-7 charset when possible. If we
need unicode support, we need to encode the messages as binary, using
UCS2 encoding. This restricts the message length to 70 characters,
often forcing messages to be split into multiple parts.

Here we attempt to encode the message in iso-8859-1. Only if this fails,
we use the UCS2 encoding.

TODO: Missing tests.

TargetProcess #5681
https://dignio.tpondemand.com/entity/5681