Fix encoding/decoding error after sms is sent
iso-8859-1 encodable messages lead to an error in Data Services after
being sent. The reason was that the SMS gateway provider endpoint
responds with a jsonified message object, and the original message text
was replaced with its encoded equivalent. This led to problems when
jsonification tried to decode the iso-8859-1 encoded string as utf-8.

TargetProcess #5793
https://dignio.tpondemand.com/entity/5793